Add nice error message support
This eliminates the System.OpenCL.Wrappers.Errors module and moves its
values into System.OpenCL.Wrappers.Types to avoid a circular dependency.
I don't think this is a big deal, since the Types module already
contains a lot of values for the types it defines, and there aren't that
many anyway.